Extract enemy sprite
